Move Metric implementations and conversion functions to the Metric.kt source file
